+++ Source/Doppelkopf/DoppelkopfMode.h
@@ -21,8 +21,8 @@ public:
 
 	ADoppelkopfMode();
 	void ShuffleCards();
-	TArray<int32> NewDeck;
-	TArray<int32> GiveCards();
+	TArray<uint8> NewDeck;
+	TArray<uint8> GiveCards();
 
 	void Tick(float DeltaTime);
 	
@@ -33,7 +33,7 @@ private:
 
 
 	int askedForCards = 0;
-	TArray<int32> CardDeck = { 
+	TArray<uint8> CardDeck = {
 		0, 1, 2, 3 , 4, 5, // Karo: A, 9, 10, J, Q K
 		6, 7, 8, 9, 10, 11, // Kreuz
 		12, 13, 14, 15, 16, 17, // Herz
@@ -44,7 +44,7 @@ private:
 		18, 19, 20, 21, 22, 23,
 	};
 
-	int32 CurrentDeck[48] = {
+	uint8 CurrentDeck[48] = {
 		0, 1, 2, 3 , 4, 5, // Karo: A, 9, 10, J, Q K
 		6, 7, 8, 9, 10, 11, // Kreuz
 		12, 13, 14, 15, 16, 17, // Herz
